feat: move collateral pools tvl to demo prime
